# docs/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AdminAccountView" : {
        "properties" : {
          "accountId" : {
            "type" : "string"
          },
          "closedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "ChangeRoleRequest" : {
        "properties" : {
          "role" : {
            "type" : "string"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "API for the Georgian housing-review platform.",
    "title" : "Geo Housing Reviews API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/api/admin/accounts/{accountId}" : {
      "get" : {
        "operationId" : "getAccount",
        "parameters" : [ {
          "in" : "path",
          "name" : "accountId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminAccountView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-account-controller" ]
      }
    },
    "/api/admin/accounts/{accountId}/role" : {
      "patch" : {
        "operationId" : "changeRole",
        "parameters" : [ {
          "in" : "path",
          "name" : "accountId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChangeRoleRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminAccountView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-account-controller" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}